Add Engine reference to run_ctx
In order to access e.g. the Datasets from within the forward_step
this commit adds a reference to the engine to the current run_ctx.

As the problematic members are all marked private, there should
be a good understanding on what should be accessed and what not.

Messing witht the engine is always on the own risk of the user.